+++ api.md
@@ -1,0 +1,152 @@
+<!-- Generated by documentation.js. Update this documentation by updating the source code. -->
+
+### Table of Contents
+
+-   [MockServer][1]
+    -   [init][2]
+        -   [Parameters][3]
+        -   [Examples][4]
+-   [InitOptions][5]
+    -   [Properties][6]
+-   [MockRequest][7]
+    -   [on][8]
+        -   [Parameters][9]
+        -   [Examples][10]
+-   [ResponseOptions][11]
+    -   [Properties][12]
+
+## MockServer
+
+### init
+
+Init mock server and set request interception on the page. All requests not
+matching `baseAppUrl` and `baseApiUrl` and not handled using special
+registered handler will be aborted and reported to console.
+
+#### Parameters
+
+-   `page` **[Object][13]** Puppeteer's page
+-   `options` **[InitOptions][14]** init options (optional, default `{}`)
+
+#### Examples
+
+```javascript
+import puppeteer from 'puppeteer';
+import mockServer from 'pptr-mock-server';
+
+// typically your global test setup
+const browser = await puppeteer.launch();
+const page = await browser.newPage();
+const baseAppUrl = 'http://localhost';
+this.mockRequest = await mockServer.init(page, {
+  baseAppUrl,
+  baseApiUrl: baseAppUrl + '/api/'
+});
+// now you can use `this.mockRequest` in your tests
+```
+
+Returns **[Promise][15]&lt;[MockRequest][16]>** 
+
+## InitOptions
+
+Type: [Object][13]
+
+### Properties
+
+-   `baseAppUrl` **[string][17]** Base app url. By default all requests matching
+    base app url are continued.
+-   `baseApiUrl` **[string][17]** Base api url. By default all requests matching
+    base api url are responded with 200 status and empty body, but you will see a
+    warning in output.
+
+## MockRequest
+
+Class for registering mock responses. It's instance is returned by
+`mockServer.init()`
+
+### on
+
+Set expected mock response for request. There are also shortcuts `.get()`,
+`.post()`, `.put()` and `.delete()` available
+
+#### Parameters
+
+-   `method` **[string][17]** request HTTP method
+-   `endpoint` **[string][17]** request endpoint URL for request. If relative URL
+    is passed, it's considered as a request to api **relative** to baseApiUrl.
+-   `status` **[number][18]** response status code
+-   `response` **[ResponseOptions][19]** additional response options
+
+#### Examples
+
+_Handle request to relative endpoint_
+
+```javascript
+const responseConfig = {body: {result: 'ok'}};
+mockRequest.on('get', 'account', 200, responseConfig);
+```
+
+_Using shortcut method and absolute url_
+
+```javascript
+const responseConfig = {body: {result: 'not found'}};
+mockRequest.get('https://example.com/test', 404, responseConfig);
+```
+
+_Simulate request timeout_
+
+```javascript
+mockRequest.post('search', null, {abort: 'timedout', delay: 10000});
+```
+
+Returns **[MockRequest][16]** 
+
+## ResponseOptions
+
+Type: [Object][13]
+
+### Properties
+
+-   `body` **[Object][13]** response body
+-   `delay` **[number][18]?** delay response for N milliseconds
+-   `abort` **[string][17]?** abort request with supplied error code
+-   `contentType` **[string][17]?** content type. Defaults to
+    `application/json`.
+
+[1]: #mockserver
+
+[2]: #init
+
+[3]: #parameters
+
+[4]: #examples
+
+[5]: #initoptions
+
+[6]: #properties
+
+[7]: #mockrequest
+
+[8]: #on
+
+[9]: #parameters-1
+
+[10]: #examples-1
+
+[11]: #responseoptions
+
+[12]: #properties-1
+
+[13]: https://developer.mozilla.org/docs/Web/JavaScript/Reference/Global_Objects/Object
+
+[14]: #initoptions
+
+[15]: https://developer.mozilla.org/docs/Web/JavaScript/Reference/Global_Objects/Promise
+
+[16]: #mockrequest
+
+[17]: https://developer.mozilla.org/docs/Web/JavaScript/Reference/Global_Objects/String
+
+[18]: https://developer.mozilla.org/docs/Web/JavaScript/Reference/Global_Objects/Number
+
+[19]: #responseoptions
