+++ src/mock-server.js
@@ -1,11 +1,34 @@
 import handleRequest from './handle-request';
-import AddHandler from './add-handler';
+import MockRequest from './mock-request';
 
-export default function() {
-  this.init = async (
-    page,
-    {baseApiUrl, baseAppUrl, timeScaleFactor = 1} = {}
-  ) => {
+/**
+ * @class
+ * @hideconstructor
+ */
+export default function MockServer() {
+  /**
+   * Init mock server and set request interception on the page. All requests not
+   * matching `baseAppUrl` and `baseApiUrl` and not handled using special
+   * registered handler will be aborted and reported to console.
+   * @param {Object} page Puppeteer's page
+   * @param {InitOptions} options init options
+   * @return {Promise<MockRequest>}
+   * @example
+   * import puppeteer from 'puppeteer';
+   * import mockServer from 'pptr-mock-server';
+   *
+   * // typically your global test setup
+   * const browser = await puppeteer.launch();
+   * const page = await browser.newPage();
+   * const baseAppUrl = 'http://localhost';
+   * this.mockRequest = await mockServer.init(page, {
+   *   baseAppUrl,
+   *   baseApiUrl: baseAppUrl + '/api/'
+   * });
+   * // now you can use `this.mockRequest` in your tests
+   */
+  this.init = async (page, options = {}) => {
+    const {baseApiUrl, baseAppUrl, timeScaleFactor = 1} = options;
     const handlers = [];
     await page.setRequestInterception(true);
     await page.on('request', async request => {
@@ -20,6 +43,15 @@ export default function() {
       }
     });
 
-    return new AddHandler(handlers, baseApiUrl, timeScaleFactor);
+    return new MockRequest(handlers, baseApiUrl, timeScaleFactor);
   };
 }
+
+/**
+ * @typedef {Object} InitOptions
+ * @property {string} baseAppUrl Base app url. By default all requests matching
+ * base app url are continued.
+ * @property {string} baseApiUrl Base api url. By default all requests matching
+ * base api url are responded with 200 status and empty body, but you will see a
+ * warning in output.
+ */
